languages api JSON response

+++ app/Http/Controllers/LanguageController.php
@@ -10,12 +10,13 @@ class LanguageController extends Controller
 {
     public function __construct()
     {
-        $this->middleware(['permission:languages_read'])->only('index');
-        $this->middleware(['permission:languages_create'])->only('create');
-        $this->middleware(['permission:languages_update'])->only('edit');
-        $this->middleware(['permission:languages_delete'])->only('destroy');
+//        $this->middleware(['permission:languages_read'])->only('index');
+//        $this->middleware(['permission:languages_create'])->only('create');
+//        $this->middleware(['permission:languages_update'])->only('edit');
+//        $this->middleware(['permission:languages_delete'])->only('destroy');
 
     }
+
     /**
      * Display a listing of the resource.
      *
@@ -23,8 +24,8 @@ class LanguageController extends Controller
      */
     public function index()
     {
-        $languages = Language::OrderBy('created_at', 'desc')->get();
-        return view('main_settings.languages.index', compact('languages'));
+        $languages = Language::OrderBy('created_at', 'desc')->paginate(10);
+        return responseJson(1, "ok", $languages);
     }
 
     /**
@@ -40,48 +41,58 @@ class LanguageController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
-    public function store(LanguageRequest $request)
+    public function store(Request $request)
     {
+        $validator = validator($request->all(), [
+            "name" => "required",
+        ]);
+
+        if ($validator->fails()) {
+            return responseJson(0, $validator->errors()->getMessages(), "");
+        }
         try {
             $language = Language::create($request->all());
             if ($language) {
-                notify()->success("تم حفظ الاعدادات بنجاح","","bottomLeft",);
-                return redirect()->route('languages.index');
+                return responseJson(1, __('data created successfully'), $language);
+
             } else {
-                notify()->error("هناك خطأ ما يرجى المحاولة فى وقت لاحق","","bottomLeft",);
-                return redirect()->route('languages.index');
+
             }
-        } catch (\Exception $th) {
-            notify()->error(" هناك خطأ ما يرجى المحاولة فى وقت لاحق","","bottomLeft",);
-            return redirect()->route('languages.index');
+        } catch (\Exception $ex) {
+            return responseJson(0, "", $ex->getMessage());
+
         }
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
     public function show($id)
     {
-        //
+        $language = Language::find($id);
+        if (!$language) {
+            return responseJson(0, __('data not found'), '');
+        }
+        return responseJson(1, "ok", $language);
     }
 
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  int  $id
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
     public function edit($id)
     {
         $language = Language::find($id);
         if (!$language) {
-            notify()->warning("هذه الاعدادات غير موجودة","","bottomLeft",);
+            notify()->warning("هذه الاعدادات غير موجودة", "", "bottomLeft",);
             return redirect()->route('languages.index');
         }
         return view('main_settings.languages.edit', compact('language'));
@@ -90,34 +101,38 @@ class LanguageController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
+     * @param \Illuminate\Http\Request $request
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
-    public function update(LanguageRequest $request, $id)
+    public function update(Request $request, $id)
     {
+        $validator = validator($request->all(), [
+            "name" => "required",
+        ]);
+
+        if ($validator->fails()) {
+            return responseJson(0, $validator->errors()->getMessages(), "");
+        }
+
         try {
             $language = Language::find($id);
 
             if (!$language) {
-                notify()->warning("هذه الاعدادات غير موجودة ","","bottomLeft",);
-                return redirect()->route('languages.index');
+                return responseJson(0, __('data not found'), '');
             } else {
                 $language->update($request->all());
-                notify()->success("تم تعديل البيانات بنجاح","","bottomLeft",);
-
-                return redirect()->route('languages.index');
+                return responseJson(1, __('data updated successfully'), $language);
             }
         } catch (\Exception $ex) {
-            notify()->error(" هناك خطأ ما يرجى المحاولة فى وقت لاحق","","bottomLeft",);
-            return redirect()->route('languages.index');
+            return responseJson(0, "", $ex->getMessage());
         }
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
@@ -125,17 +140,16 @@ class LanguageController extends Controller
         try {
             $language = Language::find($id);
             if (!$language) {
-                notify()->warning("هذه البيانات غير موجودة","","bottomLeft",);
-
-                return redirect()->route('languages.index');
+                return responseJson(0, __('data not found'), '');
+            }
+            if ($language->applicationsLang1->count() > 0 || $language->applicationsLang2->count() > 0 || $language->studentsLang1->count() > 0 || $language->studentsLang1->count() > 0) {
+                return responseJson(0, __('this item can not be deleted'), $language->fresh());
             }
             $language->delete();
-            notify()->success("تم الحذف بنجاح","","bottomLeft",);
-            return redirect()->route('languages.index');
+            return responseJson(1, __('deleted successfully'), '');
 
         } catch (\Exception $ex) {
-            notify()->error(" هناك خطأ ما يرجى المحاولة فى وقت لاحق","","bottomLeft",);
-            return redirect()->route('languages.index');
+            return responseJson(0, "", $ex->getMessage());
         }
     }
 }
